New Lich spellbook
Liches and Ancient Liches have several iconic spells, conjurations,
hexes, and summon greater demon (espeically notable on ancient liches
who are tough enough to keep their demons around!). So many that making
a single book that covers all of the roles leaves a book that steps on
its own toes.

This commit creates a conjurations focused spellbook for Liches and
Ancient Liches consisting of Corrosive Bolt, LCS, Slow, Haste, and
Invisibility; this is basically the old LCS book, but with Slow and
Invis instead of Petrify. A subsequent commit will introduce a new 'L'
monster to take up the summoning Lich role.